+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- markdownlint-disable MD024 -->
 
+## [Unreleased]
+
+### Added
+
+- detection of hg pinned revision in `fab init`
+
+### Changed
+
+### Removed
+
 ## [2.0.0]
 
 ### Added